refactor(headless-crawler): jobDetailExtractorをE-jobDetailモジュールへ移行・整理
- 旧jobDetailExtractorのcontext/indexを削除
- E-jobDetailモジュール（context, error, index, playground）を新規追加
- handlerで新APIを利用するよう修正
- コード構成をETL/DI設計に合わせてリファクタリング
